requirements.txt
Django>=2.0.0,<3
channels>=2.0.2
channels_redis~=2.0
django-registration-redux
django-autocomplete-light
django-taggit
django-extensions
drf-access-policy
django-cors-headers
django-crispy-forms
django-filter
django-pandas
djangorestframework
drf-access-policy
django-oauth-toolkit
psycopg2
python-slugify
pyyaml
uritemplate
boto3
django-storages
graphene-django
s3fs

graphene-django-extras

zarr
blosc
xarray
lz4
numcodecs
dask[complete]==2.9.2
markdown

ptvsd
pytest-django 
pytest-asyncio
pylint
pylint-django